# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "one piece 3mk",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
...
        "/users/{user}/locations": {
            "get": {
                "tags": [
                    "Users",
                    "Locations"
                ],
                "summary": "Get locations of a user",
                "description": "Retrieve the locations (name, longitude, and latitude) for a specific user.",
                "operationId": "dc8b927ad770c7bd6b8f21fc9bb9d261",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of locations for the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "locations": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "Name of the location",
                                                        "type": "string"
                                                    },
                                                    "longitude": {
                                                        "description": "Longitude of the location",
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "latitude": {
                                                        "description": "Latitude of the location",
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized to view locations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users",
                    "Locations"
                ],
                "summary": "Add locations for a user",
                "description": "Allows the authenticated user to add locations with longitude, latitude, and name. A maximum of 5 locations can be added.",
                "operationId": "2d58885eb5fb2868b0a847e3776cb30f",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body for adding locations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locations"
                                ],
                                "properties": {
                                    "locations": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "longitude",
                                                "latitude",
                                                "name"
                                            ],
                                            "properties": {
                                                "longitude": {
                                                    "description": "Longitude of the location",
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "latitude": {
                                                    "description": "Latitude of the location",
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "name": {
                                                    "description": "Name of the location",
                                                    "type": "string",
                                                    "maxLength": 50
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "maxItems": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Locations added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Locations added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error, invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized to add locations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use 'Bearer <token>' in the Authorization header",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}